Poprawy po reciew - odczytywanie domain z .env, wydzielenie metod
