+++ .gitignore
@@ -1,1 +1,3 @@
 scrape-result/img/
+Debug/
+shop/img/
